+++ packages/studio-base/src/components/PanelCatalog/index.ts
@@ -3,3 +3,4 @@
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
 export * from "./PanelCatalog";
+export * from "./types";
